refactor: move from uint to int
I was using uint to enforce values being positive, but this is not a
good use case for the type. It also overcomplicates things due to so
much casting between uint and int. Instead I will validate the Scoring
type, which is the container of the strictly positive integers.

+++ internal/tui/menu/model.go
@@ -168,10 +168,10 @@ func (m Model) startGame() (tea.Cmd, error) {
 
 	switch mode {
 	case common.ModeMarathon:
-		in := common.NewSingleInput(mode, uint(level), playerName)
+		in := common.NewSingleInput(mode, level, playerName)
 		return common.SwitchModeCmd(mode, in), nil
 	case common.ModeUltra:
-		in := common.NewSingleInput(mode, uint(level), playerName)
+		in := common.NewSingleInput(mode, level, playerName)
 		return common.SwitchModeCmd(mode, in), nil
 	case common.ModeMenu, common.ModeLeaderboard:
 		return nil, fmt.Errorf("invalid mode for starting game: %q", mode)
